+++ sdk/ai/Azure.AI.Agents.Persistent/tests/Samples/Sample_PersistentAgents_Basics.cs
@@ -9,7 +9,6 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Azure.Core.TestFramework;
-using Azure.Identity;
 using NUnit.Framework;
 
 namespace Azure.AI.Agents.Persistent.Tests;
@@ -28,7 +27,7 @@ public partial class Sample_PersistentAgents_Basics : SamplesBase<AIAgentsTestEn
         var projectEndpoint = TestEnvironment.PROJECT_ENDPOINT;
         var modelDeploymentName = TestEnvironment.MODELDEPLOYMENTNAME;
 #endif
-        PersistentAgentsClient client = new(projectEndpoint, new AzureCliCredential());
+        PersistentAgentsClient client = new(projectEndpoint, new DefaultAzureCredential());
         #endregion
 
         // Step 1: Create an agent